requirements.txt
django
python-dotenv==0.15.0